+++ changelog.md
@@ -1,3 +1,13 @@
+Version 1.0.2
+-------------
+* Add addSearchPath.
+* Add -d:hint_glue_code switch.
+* Add addFnDoc, addMethodDoc, and other modification for better docstring support.
+* PocketLang DSL can add/modify docstring of function/class by comment statement.
+* Add `Comment Statement` section in tests/tests.nim
+* Fix bug about symbol as identifiers.
+* Fix destructors compatibility and update for Nim Compiler 2.2.
+
 Version 1.0.1
 -------------
 * Update pocketlang version.